+++ packages/admin-ui-docs/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.18.1-dev.11 (2023-06-29)
+
+**Note:** Version bump only for package @vtex/admin-ui-docs
+
+
+
+
+
+## [0.18.1-dev.10](https://github.com/vtex/admin-ui/compare/@vtex/admin-ui-docs@0.18.1-dev.9...@vtex/admin-ui-docs@0.18.1-dev.10) (2023-06-28)
+
+**Note:** Version bump only for package @vtex/admin-ui-docs
+
+
+
+
+
+## 0.18.1-dev.9 (2023-06-27)
+
+**Note:** Version bump only for package @vtex/admin-ui-docs
+
+
+
+
+
+## 0.18.1-dev.8 (2023-06-22)
+
+**Note:** Version bump only for package @vtex/admin-ui-docs
+
+
+
+
+
 ## 0.18.1-dev.7 (2023-06-19)
 
 **Note:** Version bump only for package @vtex/admin-ui-docs